requirements.txt
typing-extensions
# langchain-openai  # removed - not using OpenAI
langchain-experimental
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
python-dotenv
questionary
langchain_anthropic
langchain-google-genai
